front: replace useSelectedTrain hook by useSelectedTimetableItem
This change is necessary because we need to manipulate the timetableItem
and not just the train in useSimulationResults.

See the following commits
